Intro lecture on GPU programming

+++ GPU/ssp.cu
@@ -1,18 +1,10 @@
 
 /* 
- * The recursive algorithm for the Subset Sum Problem (SSP) on GPU
+ * Recursive algorithm for the Subset Sum Problem (SSP) on GPU
  *
  * We are only interested in verifying whether solutions exist;
  * we do not intend to output the list of integers forming the 
- * found solution (i.e. the subsets)
- *
- * Execution on Nuvolos with a Tesla T4 GPU:
- *
- * Recursive algorithm for SSP with CUDA on GPU
- * nblocks = 2048, nthreads = 128
- * SSP size = 18, search space size = 262144
- * Recursive algorithm says that 6840 subsets correspond to the SSP target (time  0.0020079999231)
- * GPU parallel approach says that 6840 subsets correspond to the SSP target (time  0.0009889999637)
+ * found solution (i.e. the subsets).
  *
  * AM
  */
@@ -25,12 +17,11 @@
 
 // every thread computes the sum of the subset encoded by the binary version
 // of its own identifier ;-)
-
-__global__ void ssp_on_gpu(size_t n,int *set_gpu,int *sum_gpu)
+__global__ void ssp_on_gpu(size_t n,unsigned long *set_gpu,unsigned long *sum_gpu)
 {
-   int sum = 0;
-   unsigned int id = (blockIdx.x*blockDim.x) + threadIdx.x;
-   unsigned int copy = id;
+   unsigned long sum = 0UL;
+   size_t id = (blockIdx.x*blockDim.x) + threadIdx.x;
+   size_t copy = id;
    for (size_t i = 0; i < n; i++)
    {
       sum = sum + (id%2)*set_gpu[i];
@@ -40,7 +31,7 @@ __global__ void ssp_on_gpu(size_t n,int *set_gpu,int *sum_gpu)
 };
 
 // recursive algorithm in sequential
-void recursive(size_t i,size_t n,int target,int partial,int *set,int *count)
+void recursive(size_t i,size_t n,unsigned long target,unsigned long partial,unsigned long *set,size_t *count)
 {
    // are we at the last element yet?
    if (i == n)
@@ -57,7 +48,7 @@ void recursive(size_t i,size_t n,int target,int partial,int *set,int *count)
    recursive(i+1,n,target,partial,set,count);
 };
 
-// computing CPU time
+// computing CPU clock time
 float compute_time(time_t start,time_t end)
 {
    return ((float)((int)end - (int)start))/CLOCKS_PER_SEC;
@@ -71,10 +62,10 @@ int main(int argc,char *argv[])
    size_t nthreads = warp_size*warps_per_block;
    size_t nblocks = 2048;  // idem
    size_t total_threads = nblocks*nthreads;
-   size_t n = 0;
-   int count,partial;
-   int target,*set,*sum;
-   int *set_gpu,*sum_gpu;
+   size_t n = 0UL;
+   unsigned long count,partial;
+   unsigned long target,*set,*sum;
+   unsigned long *set_gpu,*sum_gpu;
    time_t startclock,endclock;
    float time;
 
@@ -92,11 +83,11 @@ int main(int argc,char *argv[])
    printf("SSP size = %lu, search space size = %lu\n",n,total_threads);
 
    // memory allocation for the CPU
-   set = (int*)calloc(n,sizeof(int));
-   sum = (int*)calloc(total_threads,sizeof(int));
+   set = (unsigned long*)calloc(n,sizeof(unsigned long));
+   sum = (unsigned long*)calloc(total_threads,sizeof(unsigned long));
 
    // generating a simple SSP instance
-   target = 0;
+   target = 0UL;
    for (size_t i = 0; i < n; i++)
    {
       set[i] = rand()%10;  // integers between 0 and 9
@@ -104,22 +95,22 @@ int main(int argc,char *argv[])
    };
 
    // counting the number of solutions in sequential
-   count = 0;
-   partial = 0;
+   count = 0UL;
+   partial = 0UL;
    startclock = clock();
    recursive(0,n,target,partial,set,&count);  // recursive calls
    endclock = clock();
    time = compute_time(startclock,endclock);
 
    // printing the solution obtained in sequential
-   printf("Recursive algorithm says that %d subsets correspond to the SSP target (time %16.13lf)\n",count,time);
+   printf("Recursive algorithm says that %lu subsets correspond to the SSP target (time %16.13lf)\n",count,time);
 
    // memory allocation on GPU
-   cudaMalloc((void**)&set_gpu,n*sizeof(int));
-   cudaMalloc((void**)&sum_gpu,total_threads*sizeof(int));
+   cudaMalloc((void**)&set_gpu,n*sizeof(unsigned long));
+   cudaMalloc((void**)&sum_gpu,total_threads*sizeof(unsigned long));
 
    // moving the set of integers on GPU global memory
-   cudaMemcpy(set_gpu,set,n*sizeof(int),cudaMemcpyHostToDevice);
+   cudaMemcpy(set_gpu,set,n*sizeof(unsigned long),cudaMemcpyHostToDevice);
 
    // launching the kernel
    startclock = clock();
@@ -127,10 +118,10 @@ int main(int argc,char *argv[])
    cudaDeviceSynchronize();
 
    // retrieving the computed sums
-   cudaMemcpy(sum,sum_gpu,total_threads*sizeof(int),cudaMemcpyDeviceToHost);
+   cudaMemcpy(sum,sum_gpu,total_threads*sizeof(unsigned long),cudaMemcpyDeviceToHost);
 
    // counting the number of sums that are equal to the target
-   count = 0;
+   count = 0UL;
    for (size_t i = 0; i < total_threads; i++)
    {
       if (sum[i] == target)  count++;
@@ -139,7 +130,7 @@ int main(int argc,char *argv[])
    // printing the result obtained with the help of the GPU
    endclock = clock();
    time = compute_time(startclock,endclock);
-   printf("GPU parallel approach says that %d subsets correspond to the SSP target (time %16.13f)\n",count,time);
+   printf("GPU parallel approach says that %lu subsets correspond to the SSP target (time %16.13f)\n",count,time);
 
    // freing memory
    cudaFree(set_gpu);
